FSPT-1370: move existing reports routes into collections
Separate out reports (list reports) from collection pages

Note this is split into a number of commits to make sure its tracked as
a `mv` and not replacing 5,000 loc.

We'll allow the collection pages to work for any collection types
for the time being, while allowing the individual pages for monitoring
and pre-award to become more domain specific as required.
