+++ app/deliver_grant_funding/routes/__init__.py
@@ -4,12 +4,12 @@ deliver_grant_funding_blueprint = Blueprint(name="deliver_grant_funding", import
 
 from app.deliver_grant_funding.routes import (  # noqa: E402, F401
     api,
+    collections,
     grant_details,
     grant_setup,
     grant_team,
     misc,
     pre_award,
-    reports,
     runner,
 )
