+++ tests/integration/deliver_grant_funding/routes/test_collections.py
@@ -75,7 +75,7 @@ from app.deliver_grant_funding.forms import (
     ReopenSubmissionForm,
     SetUpReportForm,
 )
-from app.deliver_grant_funding.routes.reports import (
+from app.deliver_grant_funding.routes.collections import (
     _determine_return_url_and_update_session_after_choosing_reference_for_expression,
 )
 from app.deliver_grant_funding.session_models import (
